Moves text-to-speech code to section, so users can hear 1 question at a time
