--- client/src/components/listen.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const Listen = () => {
-    // TODO: Access API here
-  return (
-    <div>
-        <h2>Listen is present</h2>
-        <button>Click me to have the lesson read to you</button>
-    </div>
-  );
-};
-
-export default Listen;
